rfr: fix broken tests and refactor of cancel action
TRACEFOSS-800
