+++ src/app/modules/shared/modules/notification/notification.module.ts
@@ -21,7 +21,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ModalModule } from '@shared/modules/modal/modal.module';
 import { ApproveNotificationModalComponent } from '@shared/modules/notification/modal/approve/approve-notification-modal.component';
-import { CancelNotificationModalComponent } from '@shared/modules/notification/modal/delete/cancel-notification-modal.component';
+import { CancelNotificationModalComponent } from '@shared/modules/notification/modal/cancel/cancel-notification-modal.component';
 import { NotificationTabComponent } from '@shared/modules/notification/notification-tab/notification-tab.component';
 import { SharedModule } from '@shared/shared.module';
 import { TemplateModule } from '@shared/template.module';
